Fixed an oversight where the type should've been float, not int, for the z values.

+++ src/Renderer.h
@@ -99,7 +99,7 @@ class Renderer {
       for (int x = box.ll.x; x <= box.ur.x; ++x) {
         auto bc_screen = barycentric(tri, glm::ivec2(x, y));
         if (bc_screen.x >= 0 && bc_screen.y >= 0 && bc_screen.z >= 0) {
-          int z = tri.a.z * bc_screen.x + tri.b.z * bc_screen.y + tri.c.z * bc_screen.z;
+          float z = tri.a.z * bc_screen.x + tri.b.z * bc_screen.y + tri.c.z * bc_screen.z;
           if (zbuffer[x][y] < z) {
             zbuffer[x][y] = z;
             frame.put_pixel(x, y, color);
